+++ ftpsync/ftp_target.py
@@ -9,12 +9,14 @@ import ftplib
 import json
 import os
 import time
-from posixpath import join as join_url, normpath as normpath_url, relpath as relpath_url
+from posixpath import join as join_url
+from posixpath import normpath as normpath_url
+from posixpath import relpath as relpath_url
 from tempfile import SpooledTemporaryFile
 
 from ftpsync.metadata import DirMetadata, IncompatibleMetadataVersion
 from ftpsync.resources import DirectoryEntry, FileEntry
-from ftpsync.targets import _Target, _get_encoding_opt
+from ftpsync.targets import _get_encoding_opt, _Target
 from ftpsync.util import (
     get_credentials_for_url,
     is_native,
@@ -366,7 +368,7 @@ class FTPTarget(_Target):
         if not path.startswith(self.root_dir):
             # paranoic check to prevent that our sync tool goes berserk
             raise RuntimeError(
-                "Tried to navigate outside root %r: %r" % (self.root_dir, path)
+                "Tried to navigate outside root {!r}: {!r}".format(self.root_dir, path)
             )
         self.ftp.cwd(dir_name)
         self.cur_dir = path
@@ -714,7 +716,7 @@ class FTPTarget(_Target):
         Returns:
             None
         """
-        LF = b"\n"
+        LF = b"\n"  # noqa N806
         buffer = b""
 
         # needed to access buffer accross function scope